Do not use window global in HMR runtime (#529)
The name of the global object may differ based on the context in which the
script is running. For example, web workers use "self" instead of "window". As
location is a property of the global, it can be referenced directly as a global
variable, which avoids having to deal with this naming inconsistency.

Fixes #528